"Double => decimal; bug fixes, restructured transaction view"

+++ Open/Data/Bank/AccountData.cs
@@ -4,11 +4,11 @@ namespace Open.Data.Bank {
     public class AccountData : IdentifiedData {
         private string type;
         private string status;
-        private double? balance;
+        private decimal? balance;
         private string applicationUserId;
 
-        public double? Balance {
-            get => getDouble(ref balance);
+        public decimal? Balance {
+            get => getDecimal(ref balance);
             set => balance = value;
         }
